HyperSpectra: fix coloring by categoric variables with missing values
Before, the levels for the color palette were set according to the
displayed image values. For selections of the data, some values can be
missing. If missing values were on the edges of variable.values,
the colors were wrong. Color palette levels for categoric variables
are now manually set to span all possible values.