fix: Fixing the "Avoid magic word" problem in the owfft data for the .gsf format
Do so by making the reader calculate its datapoint spacing and save it as
metadata that can be consumed in the fft widget. This allow the owfft widget to
treat the data from the NeaReaderGSF and the NeaReaderMultiChannel the same
way, as it should since they both use the same complex fft and provide data in
similar form.